requirements.txt
# LangChain 核心套件
langchain
langchain-core
langchain-openai
langchain-community
langchain-text-splitters

# PDF 處理
pypdf

# Google Gemini 整合
langchain-google-genai

# 其他相依套件
python-dotenv>=0.21.0
faiss-cpu
numexpr

# 如果需要在 Jupyter Notebook 中使用
# jupyter  

# 如果需要建立 Web 介面
# streamlit  